+++ gym_gazebo/__init__.py
@@ -4,6 +4,6 @@ from gym.envs.registration import register
 # F1 envs
 register(
     id='F1Env-v0',
-    entry_point='gym_gazebo.envs.f1.modes:F1Env',
+    entry_point='gym_gazebo.envs.f1.models:F1Env',
     # More arguments here
 )
