+++ gym_gazebo/envs/f1/env_type.py
@@ -2,6 +2,7 @@ from enum import Enum
 
 
 class TrainingType(Enum):
-    qlearn_env = "qlearn"
+    qlearn_env_camera = "qlearn_camera"
+    qlearn_env_laser = "qlearn_laser"
     dqn_env = "dqn"
     manual_env = "manual"
