+++ gym_gazebo/envs/f1/models/__init__.py
@@ -1,6 +1,7 @@
 from gym_gazebo.envs.f1.env_type import TrainingType
 from gym_gazebo.envs.f1.exceptions import NoValidTrainingType
 
+
 class F1Env:
 
     def __new__(cls, **config):
@@ -16,16 +17,20 @@ class F1Env:
 
         training_type = config.get("training_type")
 
-        if training_type == TrainingType.qlearn_env.value:
-            from gym_gazebo.envs.f1.modes.f1_env_qlearn_camera import F1QlearnCameraEnv
+        if training_type == TrainingType.qlearn_env_camera.value:
+            from gym_gazebo.envs.f1.models.f1_env_qlearn_camera import F1QlearnCameraEnv
             return F1QlearnCameraEnv(**config)
 
+        elif training_type == TrainingType.qlearn_env_laser.value:
+            from gym_gazebo.envs.f1.models.f1_env_qlearn_laser import F1QlearnLaserEnv
+            return F1QlearnLaserEnv(**config)
+
         elif training_type == TrainingType.dqn_env.value:
-            from gym_gazebo.envs.f1.modes.f1_env_dqn_camera import GazeboF1CameraEnvDQN
+            from gym_gazebo.envs.f1.models.f1_env_dqn_camera import GazeboF1CameraEnvDQN
             return GazeboF1CameraEnvDQN(**config)
 
         elif training_type == TrainingType.manual_env.value:
-            from gym_gazebo.envs.f1.modes.f1_env_manual_pilot import GazeboF1ManualCameraEnv
+            from gym_gazebo.envs.f1.models.f1_env_manual_pilot import GazeboF1ManualCameraEnv
             return GazeboF1ManualCameraEnv(**config)
 
         else:
